#29 랜덤 문자 꺼내기 개선 - jest 추가
