+++ backend/package.json
@@ -9,11 +9,13 @@
     "@types/axios": "^0.14.0",
     "@types/dotenv": "^6.1.1",
     "@types/express": "^4.17.0",
+    "@types/jest": "^24.0.17",
     "@types/node": "^12.0.10",
-    "babel-runtime": "^6.26.0",
     "gql-merge": "^0.0.6",
     "graphql-to-typescript": "^0.1.2",
+    "jest": "^24.8.0",
     "nodemon": "^1.19.1",
+    "ts-jest": "^24.0.2",
     "ts-node": "^8.3.0",
     "tslint-config-prettier": "^1.18.0",
     "typescript": "^3.5.2"
@@ -21,7 +23,8 @@
   "scripts": {
     "dev": "cd src && nodemon --exec ts-node index.ts -e ts,graphql",
     "pretypes": "gql-merge --out-file ./src/schema.graphql ./src/api/**/*.graphql ./src/api/*/*/*.graphql",
-    "types": "graphql-to-typescript ./src/schema.graphql ./src/types/graph.d.ts"
+    "types": "graphql-to-typescript ./src/schema.graphql ./src/types/graph.d.ts",
+    "test": "jest"
   },
   "dependencies": {
     "app-root-path": "^2.2.1",
